+++ src/z3c/autoinclude/api.py
@@ -2,6 +2,7 @@ import os
 
 DEP_KEY = 'Z3C_AUTOINCLUDE_DEPENDENCIES_DISABLED'
 PLUGIN_KEY = 'Z3C_AUTOINCLUDE_PLUGINS_DISABLED'
+DEBUG_KEY = 'Z3C_AUTOINCLUDE_DEBUG'
 
 
 def dependencies_disabled():
@@ -13,7 +14,7 @@ def disable_dependencies():
 
 
 def enable_dependencies():
-    del os.environ[DEP_KEY]
+    os.environ.pop(DEP_KEY, None)
 
 
 def plugins_disabled():
@@ -25,4 +26,16 @@ def disable_plugins():
 
 
 def enable_plugins():
-    del os.environ[PLUGIN_KEY]
+    os.environ.pop(PLUGIN_KEY, None)
+
+
+def debug_enabled():
+    return DEBUG_KEY in os.environ
+
+
+def disable_debug():
+    os.environ.pop(DEBUG_KEY, None)
+
+
+def enable_debug():
+    os.environ[DEBUG_KEY] = 'True'